Refactor singletons to standard classes

+++ src/main/java/core/basesyntax/service/impl/PurchaseTransactionHandler.java
@@ -6,17 +6,11 @@ import core.basesyntax.service.TransactionHandler;
 import java.math.BigInteger;
 
 public class PurchaseTransactionHandler implements TransactionHandler {
-    private static final TransactionHandler INSTANCE = new PurchaseTransactionHandler();
-
-    private PurchaseTransactionHandler() {
-    }
-
-    public static TransactionHandler instance() {
-        return INSTANCE;
-    }
-
     @Override
     public void handle(Database database, FruitType fruitType, BigInteger quantity) {
+        if (quantity == null || quantity.signum() < 0) {
+            throw new IllegalArgumentException("Quantity is wrong. Cannot be: " + quantity);
+        }
         database.remove(fruitType, quantity);
     }
 }
